requirements.txt
langchain
openai
python-dotenv
pandas
docx2txt
PyMuPDF
